requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
bcrypt==4.0.1
gunicorn==20.1.0